refactor: Split `MenuAction` from `app.rs`
Bump cargo dependencies.
